requirements.txt
# Core dependencies
feedparser>=6.0.10,<7
Jinja2>=3.1.2,<4
PyYAML>=6.0.1,<7
requests>=2.31.0,<3
python-dateutil>=2.8.2,<3
python-dotenv>=1.0.0,<2
beautifulsoup4>=4.12.3,<5
lxml>=5.2.2,<6

# Translation via OpenAI GPT-4o-mini
openai>=1.0.0

# Optional: free translation fallback (unreliable)
# deep-translator>=1.11.0
